+++ src/DatePicker/DatePicker.js
@@ -265,7 +265,7 @@ class DatePicker extends Component {
         this.props.onFocus(this.getCallbackData());
     }
 
-    updateDate = (date) => {
+    updateDate = (date, forceStayOpen) => {
         let closeCalendar = false;
         const { formattedDate } = this.state;
 
@@ -299,7 +299,7 @@ class DatePicker extends Component {
             });
         }
 
-        if (closeCalendar) {
+        if (!forceStayOpen && closeCalendar) {
             const popover = this.popoverRef && this.popoverRef.current;
             popover && popover.handleEscapeKey();
             this.setState({ isExpanded: false }, ()=> {
@@ -357,8 +357,10 @@ class DatePicker extends Component {
             localizedText,
             onBlur,
             onDatePickerClose,
+            openToDate,
             popoverProps,
             readOnly,
+            showToday,
             specialDays,
             todayAction,
             validationState,
@@ -423,7 +425,9 @@ class DatePicker extends Component {
                                 locale={locale}
                                 localizedText={localizedText}
                                 onChange={this.updateDate}
+                                openToDate={openToDate}
                                 ref={this.calendarRef}
+                                showToday={showToday}
                                 specialDays={specialDays}
                                 weekdayStart={weekdayStart} />
                             { this._showTodayFooter() &&