+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.11.0-rc.15](https://github.com/SAP/fundamental-react/compare/v0.11.0-rc.14...v0.11.0-rc.15) (2020-06-19)
+
+
+### Bug Fixes
+
+* add polyfills for ie11 issues ([#1099](https://github.com/SAP/fundamental-react/issues/1099)) ([7c5bae2](https://github.com/SAP/fundamental-react/commit/7c5bae23f2f48e1238e590a25c2d9513a52d27e3))
+
+## [0.11.0-rc.14](https://github.com/SAP/fundamental-react/compare/v0.11.0-rc.13...v0.11.0-rc.14) (2020-06-19)
+
+
+### Features
+
+* openToDate prop for Calendar ([#1096](https://github.com/SAP/fundamental-react/issues/1096)) ([df9399e](https://github.com/SAP/fundamental-react/commit/df9399efe7f937430bc5ba33357b2e7370573706))
+
+## [0.11.0-rc.13](https://github.com/SAP/fundamental-react/compare/v0.11.0-rc.12...v0.11.0-rc.13) (2020-06-19)
+
+
+### Features
+
+* update fundamental-styles to 0.11.0-rc.34 ([#1094](https://github.com/SAP/fundamental-react/issues/1094)) ([bb059aa](https://github.com/SAP/fundamental-react/commit/bb059aaa562a5d2637a3db04a6d48cd20bc1b6f5))
+
+## [0.11.0-rc.12](https://github.com/SAP/fundamental-react/compare/v0.11.0-rc.11...v0.11.0-rc.12) (2020-06-19)
+
+
+### Bug Fixes
+
+* style sidebar ([#1097](https://github.com/SAP/fundamental-react/issues/1097)) ([1ff2bfb](https://github.com/SAP/fundamental-react/commit/1ff2bfb627f39335cf1d032584ad951cfde4b5ac))
+
 ## [0.11.0-rc.11](https://github.com/SAP/fundamental-react/compare/v0.11.0-rc.10...v0.11.0-rc.11) (2020-06-15)
 
 ## [0.11.0-rc.10](https://github.com/SAP/fundamental-react/compare/v0.11.0-rc.9...v0.11.0-rc.10) (2020-06-12)